Rename to FormFieldInputContainer
