--- packages/form-field/src/FormFieldInput/FormFieldInput.types.ts
@@ -1,8 +1,0 @@
-import { HTMLElementProps } from '@leafygreen-ui/lib';
-
-import { FormFieldChildren } from '../FormField/FormField.types';
-
-export interface FormFieldInputProps extends HTMLElementProps<'div'> {
-  children: FormFieldChildren;
-  icon?: React.ReactElement;
-}
