--- packages/form-field/src/FormFieldInput/index.ts
@@ -1,2 +1,0 @@
-export { FormFieldInput } from './FormFieldInput';
-export { FormFieldInputProps } from './FormFieldInput.types';
