+++ packages/form-field/src/FormFieldInputContainer/FormFieldInputContainer.types.ts
@@ -1,0 +1,15 @@
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+
+import { FormFieldChildren } from '../FormField/FormField.types';
+
+export interface FormFieldInputContainerProps extends HTMLElementProps<'div'> {
+  /**
+   * Must pass a single element in order to label the input element appropriately
+   */
+  children: FormFieldChildren;
+
+  /**
+   * The icon rendered on the right of the input box
+   */
+  icon?: React.ReactElement;
+}
