+++ packages/form-field/src/FormFieldInputContainer/index.ts
@@ -1,0 +1,2 @@
+export { FormFieldInputContainer } from './FormFieldInputContainer';
+export { FormFieldInputContainerProps } from './FormFieldInputContainer.types';
